Rename Setup.js to Admin.js
